Move InvTweaksContainerSection to api package.
