+++ src/minecraft/invtweaks/InvTweaksHandlerAutoRefill.java
@@ -1,43 +1,39 @@
 package invtweaks;
 
-import invtweaks.InvTweaksConst;
-import invtweaks.InvTweaksItemTree;
-import invtweaks.InvTweaksItemTreeItem;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeoutException;
 import java.util.logging.Logger;
 
-
+import invtweaks.api.InvTweaksContainerSection;
 import net.minecraft.client.Minecraft;
 import net.minecraft.src.InvTweaksObfuscation;
 import net.minecraft.src.ItemStack;
 
 /**
  * Handles the auto-refilling of the hotbar.
- * 
+ *
  * @author Jimeo Wan
  *
  */
 public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
-    
+
     private static final Logger log = Logger.getLogger("InvTweaks");
 
     private InvTweaksConfig config = null;
-    
+
     public InvTweaksHandlerAutoRefill(Minecraft mc, InvTweaksConfig config) {
 		super(mc);
 		setConfig(config);
 	}
-    
+
     public void setConfig(InvTweaksConfig config) {
     	this.config = config;
     }
-    
+
 	/**
      * Auto-refill
-	 * @throws Exception 
+	 * @throws Exception
      */
 	public void autoRefillSlot(int slot, int wantedId, int wantedDamage) throws Exception {
 
@@ -47,19 +43,19 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
 		ItemStack candidateStack, replacementStack = null;
 		int replacementStackSlot = -1;
         boolean refillBeforeBreak = config.getProperty(InvTweaksConfig.PROP_AUTO_REFILL_BEFORE_BREAK)
-                .equals(InvTweaksConfig.VALUE_TRUE); 
-		
+                .equals(InvTweaksConfig.VALUE_TRUE);
+
 		List<InvTweaksConfigSortingRule> matchingRules = new ArrayList<InvTweaksConfigSortingRule>();
 		List<InvTweaksConfigSortingRule> rules = config.getRules();
 		InvTweaksItemTree tree = config.getTree();
-		
+
 		// Check that the item is in the tree
         if (!tree.isItemUnknown(wantedId, wantedDamage)) {
-            
+
             //// Search replacement
-            
+
     		List<InvTweaksItemTreeItem> items = tree.getItems(wantedId, wantedDamage);
-    
+
     		// Find rules that match the slot
     		for (InvTweaksItemTreeItem item : items) {
     			// Since we search a matching item using rules,
@@ -69,7 +65,7 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
     					InvTweaksConst.INVENTORY_SIZE, InvTweaksConst.INVENTORY_ROW_SIZE));
     		}
     		for (InvTweaksConfigSortingRule rule : rules) {
-    			if (rule.getType() == InvTweaksConfigSortingRuleType.SLOT 
+    			if (rule.getType() == InvTweaksConfigSortingRuleType.SLOT
     			        || rule.getType() == InvTweaksConfigSortingRuleType.COLUMN) {
     				for (int preferredSlot : rule.getPreferredSlots()) {
     					if (slot == preferredSlot) {
@@ -79,7 +75,7 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
     				}
     			}
     		}
-    
+
     		// Look only for a matching stack
     		// First, look for the same item,
     		// else one that matches the slot's rules
@@ -110,12 +106,12 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
     			}
     		}
         }
-        
+
         // If item is unknown, look for exact same item
         else {
             for (int i = 0; i < InvTweaksConst.INVENTORY_SIZE; i++) {
                 candidateStack = container.getItemStack(i);
-                if (candidateStack != null && 
+                if (candidateStack != null &&
                         getItemID(candidateStack) == wantedId &&
                         getItemDamage(candidateStack) == wantedDamage) {
                     replacementStack = candidateStack;
@@ -124,15 +120,15 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
                 }
             }
         }
-		
+
 		//// Proceed to replacement
-	
+
 		if (replacementStack != null || (refillBeforeBreak && getStack(container.getSlot(slot)) != null)) {
-			
+
 			log.info("Automatic stack replacement.");
-			
+
 		    /*
-		     * This allows to have a short feedback 
+		     * This allows to have a short feedback
 		     * that the stack/tool is empty/broken.
 		     */
 			new Thread(new Runnable() {
@@ -141,7 +137,7 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
 				private int targetedSlot;
 				private int i, expectedItemId;
 				private boolean refillBeforeBreak;
-				
+
 				public Runnable init(Minecraft mc,
 						int i, int currentItem, boolean refillBeforeBreak) throws Exception {
 					this.containerMgr = new InvTweaksContainerSectionManager(
@@ -158,9 +154,9 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
 					this.refillBeforeBreak = refillBeforeBreak;
 					return this;
 				}
-				
+
 				public void run() {
-					
+
 					// Wait for the server to confirm that the
 					// slot is now empty
 					int pollingTime = 0;
@@ -176,7 +172,7 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
 						trySleep(InvTweaksConst.AUTO_REFILL_DELAY - pollingTime);
 					if (pollingTime >= InvTweaksConst.POLLING_TIMEOUT)
 						log.warning("Autoreplace timout");
-					
+
 					// In POLLING_DELAY ms, things might have changed
 					try {
 						ItemStack stack = containerMgr.getItemStack(i);
@@ -202,19 +198,19 @@ public class InvTweaksHandlerAutoRefill extends InvTweaksObfuscation {
 						}
 					}
 					catch (NullPointerException e) {
-						// Nothing: Due to multithreading + 
+						// Nothing: Due to multithreading +
 						// unsafe accesses, NPE may (very rarely) occur (?).
 					} catch (TimeoutException e) {
 						log.severe("Failed to trigger autoreplace: "+e.getMessage());
 					}
-					
+
 				}
-				
+
 			}.init(mc, replacementStackSlot, slot, refillBeforeBreak)).start();
-			
+
 		}
     }
-	
+
 	private static void trySleep(int delay) {
 		try {
 			Thread.sleep(delay);
